Change GET /solve/ to GET /solution/
